fix some

+++ app/backboard/views.py
@@ -26,6 +26,9 @@ def logout(request):
     return redirect('/backboard/')
 
 def all_cases(request):
+    if not request.user.is_authenticated or not request.user.is_staff:
+        return redirect('/backboard/')
+
     cases = Case.objects.all()
     state = request.GET.get('state')
     if state != None:
@@ -41,6 +44,9 @@ def all_cases(request):
     return render(request, 'backboard/all_cases.html',{'cases':page_obj})
 
 def all_members(request):
+    if not request.user.is_authenticated or not request.user.is_staff:
+        return redirect('/backboard/')
+
     users = User.objects.filter(is_staff=False)
     paginator = Paginator(users, 10)
     if request.GET.get('page') != None:
@@ -53,6 +59,9 @@ def all_members(request):
     return render(request, 'backboard/all_members.html',{'users':page_obj})
 
 def bills(request):
+    if not request.user.is_authenticated or not request.user.is_staff:
+        return redirect('/backboard/')
+    
     summarys = MonthSummary.objects.all().order_by('-id')[:2]
     this_month_day = summarys[0].month_date
     last_month_day = this_month_day - datetime.timedelta(days=30)
@@ -60,6 +69,9 @@ def bills(request):
     return render(request, 'backboard/bills.html', {'summarys':summarys,  'this_month_day':this_month_day, 'last_month_day':last_month_day})
 
 def case_detail(request):
+    if not request.user.is_authenticated or not request.user.is_staff:
+        return redirect('/backboard/')
+
     case_id = request.GET.get('case')
     case = Case.objects.get(id=case_id)
     order = Order.objects.get(case=case)
@@ -69,6 +81,9 @@ def case_detail(request):
     return render(request, 'backboard/case_detail.html',{'order_increase_services':order_increase_services, 'case':case,'review':review,'order':order, 'order_increase_services':order_increase_services})
 
 def member_detail(request):
+    if not request.user.is_authenticated or not request.user.is_staff:
+        return redirect('/backboard/')
+
     user_id = request.GET.get('user')
     user = User.objects.get(id=user_id)
     offend_orders = Order.objects.filter(user=user)
@@ -76,6 +91,9 @@ def member_detail(request):
     return render(request, 'backboard/member_detail.html',{'user':user,'offend_orders':offend_orders,'take_orders':take_orders})
 
 def all_blogs(request):
+    if not request.user.is_authenticated or not request.user.is_staff:
+        return redirect('/backboard/')
+
     if request.GET.get('delete_id') != None:
         BlogPost.objects.get(id=request.GET.get('delete_id')).delete()
 
@@ -94,6 +112,9 @@ def all_blogs(request):
 
 # edit 跟 new 同一頁
 def new_blog(request):
+    if not request.user.is_authenticated or not request.user.is_staff:
+        return redirect('/backboard/')
+
     categories = BlogCategory.objects.all()
 
     if request.method == 'POST':
@@ -142,6 +163,9 @@ def new_blog(request):
     return render(request, 'backboard/new_blog.html', {'categories':categories, 'form':form})
 
 def all_categories(request):
+    if not request.user.is_authenticated or not request.user.is_staff:
+        return redirect('/backboard/')
+
     if request.GET.get('delete_id') != None:
         try:
             BlogCategory.objects.get(id=request.GET.get('delete_id')).delete()
@@ -152,6 +176,8 @@ def all_categories(request):
     return render(request, 'backboard/all_categories.html', {'categories':categories})
 
 def new_edit_category(request):
+    if not request.user.is_authenticated or not request.user.is_staff:
+        return redirect('/backboard/')
 
     if request.method == 'POST':
         if request.POST.get('post') == 'save':
@@ -171,6 +197,9 @@ def new_edit_category(request):
     return render(request, 'backboard/new_edit_category.html')
 
 def member_data_review(request):
+    if not request.user.is_authenticated or not request.user.is_staff:
+        return redirect('/backboard/')
+
     user_id = request.GET.get('user')
     user = User.objects.get(id=user_id)
     licences = License.objects.all().order_by('id')
@@ -195,6 +224,9 @@ def member_data_review(request):
     return render(request, 'backboard/member_data_review.html',{'user':user,'userLicenseImages':userLicenseImages})
 
 def refunds(request):
+    if not request.user.is_authenticated or not request.user.is_staff:
+        return redirect('/backboard/')
+
     case_id = request.GET.get('case')
     case = Case.objects.get(id=case_id)
     order = Order.objects.get(case=case)
